requirements.txt
torch
torchvision
numpy
pillow
tqdm
fastapi
uvicorn
python-multipart
